+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # News
 
+## v0.10.18 - 2025-06-30
+
+- `IGraphAlg` introduced as a generic dispatch argument for `Graphs.jl` functions extended with implementations from `igraph`.
+- Implementing `Graphs.radius`, `Graphs.diameter`, and `Graphs.Experimental.has_isomorph` methods through `igraph`.
+- `igraphalg_methods` lists all the `Graphs.jl` functions for which an `IGraphAlg` method has been defined.
+
 ## v0.10.17 - 2025-06-29
 
 - `IGNull` is introduced as a convenient placehold argument for when the low-level C function expects a `NULL` as a default.